backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.8.0
pydantic-settings==2.1.0
pydantic[email]==2.8.0
motor==3.3.2
pymongo==4.16.0
python-dotenv==1.0.0
aiohttp==3.10.1
websockets==12.0
python-multipart==0.0.6
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7
httpx==0.25.1
beautifulsoup4==4.12.2
lxml==5.1.0
email-validator==2.1.0